refactor : seperate ArticleWrite component from ArticleWritePage - reason : to seperate CSR from SSR
