+++ src/app/article/[id]/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from 'next/navigation'
 import { getArticleById, GetArticleResponseInterface } from '@/apis/articles'
 import ArticleRead from '@/containers/Article/Read'
 
-const ArticlePage = async ({ params }: { params: { id: string } }) => {
+const ArticleReadPage = async ({ params }: { params: { id: string } }) => {
   const { id } = params
 
   const loadedArticle = async (): Promise<
@@ -28,4 +28,4 @@ const ArticlePage = async ({ params }: { params: { id: string } }) => {
 
   return <ArticleRead article={article} />
 }
-export default ArticlePage
+export default ArticleReadPage
